test: move openNeovimWithNvimYaziPlugin helper to utils file
